refactor(search): Move isSameQuery to utils, fix effectiveQuery derivation
- Move isSameQuery (and its eqEq helper) from SearchTracePage/url.ts to
  utils/search-query.ts to remove the UI-layer import from useTraceDiscovery.
  url.ts re-exports it to preserve existing import paths.

- Derive effectiveQuery from queryKey[1] rather than state.data?.query so
  the entry is found even while still fetching (data is undefined until the
  first successful response).

- Replace setTimeout(10) in the test with vi.setSystemTime() for
  deterministic dataUpdatedAt ordering without flakiness risk.
